# random: Optimize `Randomizer::getBytesFromString()`