--- a/ext/random/randomizer.c
+++ b/ext/random/randomizer.c
@@ -432,6 +432,8 @@ PHP_METHOD(Random_Randomizer, getBytesFromString)
 		mask |= mask >> 2;
 		mask |= mask >> 4;
 
+		/* Example: if mask is 0xAB, will be 0xABABABABABABABAB */
+		uint64_t mask_repeat = (~((uint64_t) 0) / 0xff) * mask;
 		int failures = 0;
 		while (total_size < length) {
 			php_random_result result = engine.algo->generate(engine.state);
@@ -440,8 +442,9 @@ PHP_METHOD(Random_Randomizer, getBytesFromString)
 				RETURN_THROWS();
 			}
 
+			uint64_t offsets = result.result & mask_repeat;
 			for (size_t i = 0; i < result.size; i++) {
-				uint64_t offset = (result.result >> (i * 8)) & mask;
+				uint64_t offset = offsets >> (i * 8) & 0xff;
 
 				if (offset > max_offset) {
 					if (++failures > PHP_RANDOM_RANGE_ATTEMPTS) {