@SakiTakamachi 2024-07-11
@Girgias 

Thank you for confirmation.

I wanted to keep this as simple as possible, but all of these changes are related to improving performance.

If you don't mind a longer explanation, I can break down the commits into smaller chunks, measure them step by step, and make it clear how performance improves.

Incidentally, the most important of these changes is that the bit mask is now calculated 8 bytes at a time, rather than 1 byte at a time.

However, other changes also have a measurable effect on performance.