@Girgias 2024-07-11
I would prefer having the commits split and indicate each performance benefit it brings, so we can decide on a case by case if the tradeoff is worth it :)