@SakiTakamachi 2024-07-11
@Girgias 
I'm very embarrassed, but it seems like I made a mistake in my measurements.
When I broke the changes down into smaller pieces, some changes didn't make sense, and one change actually slowed me down.

I kept only the changes that really worked and reverted the rest.

The measurement results for commits 1 and 2 are listed in the explanation. I also re-measured "before".