@Girgias 2024-07-12
No worries, it happens :)

Now I can definitely see why the change improved the performance even without looking at the resulting assembly!

I'll wait for @TimWolla to approve the PR.