@TimWolla 2024-07-19
I'm afraid I'm unable to reproduce the improvements to the degree that your initial post indicates. I'm seeing a 1% difference between df6d85acf852e085318293b1bcbf9f9d384e5bea and the latest commit in this PR.

I am using a Intel(R) Core(TM) i7-1365U and I am compiling with:

```
./configure --enable-zend-test --enable-option-checking=fatal --enable-phpdbg --enable-fpm --enable-werror CC=clang-16 CXX=clang++-16
```

My test script is:

```
<?php
use Random\Randomizer;
use Random\Engine\PcgOneseq128XslRr64;
$r = new Randomizer(new PcgOneseq128XslRr64(0));
// 256 characters
$str = '1GCOFExQwWVNBFxOpGXilQdjcCrS6CxJRbgRf214G6w0bnFdwlmylDIpAQlcHVH5co5heIrarVoodaGeTbMQwWaM16fwXfLr03tjYymp2hg0XI5MqTlEr1taXFyHD5fYNqX7nkFJKOYU3FRMeIUkXAMpqqkIdNXP2od5Wbkra5rvLV4WT1lGN0Yg0AeEXAPq8nRxdePU2M6vQOX7wvJKKrmMAWsopEEDAdOxtrOb8lp7oPI3RjhXMbTFXaPgjljO';

for ($i = 0; $i < 100000000; $i++) {
    $r->getBytesFromString($str, 16);
}
```
(using a fixed seed to ensure that the seeding does not have an impact)

and then running the benchmark using:

```
hyperfine 'sapi/cli/php -d zend_extension=php-src/modules/opcache.so -d opcache.enable_cli=1 test.php' \
  '/tmp/unoptimized -d zend_extension=php-src/modules/opcache.so -d opcache.enable_cli=1 test.php' \
  '/tmp/unoptimized-commit1 -d zend_extension=php-src/modules/opcache.so -d opcache.enable_cli=1 test.php'

```

to get hyperfine to make the comparison for me instead of needing to manually compare the numbers.

`/tmp/unoptimized` is the version in df6d85acf852e085318293b1bcbf9f9d384e5bea, `/tmp/unoptimized-commit1` is the first commit of this PR and `sapi/cli/php` is the PR.

My results are:

```
Benchmark 1: sapi/cli/php -d zend_extension=php-src/modules/opcache.so -d opcache.enable_cli=1 test.php
  Time (mean ± σ):      3.207 s ±  0.054 s    [User: 3.203 s, System: 0.003 s]
  Range (min … max):    3.141 s …  3.310 s    10 runs
 
Benchmark 2: /tmp/unoptimized -d zend_extension=php-src/modules/opcache.so -d opcache.enable_cli=1 test.php
  Time (mean ± σ):      3.215 s ±  0.029 s    [User: 3.213 s, System: 0.002 s]
  Range (min … max):    3.169 s …  3.255 s    10 runs
 
Benchmark 3: /tmp/unoptimized-commit1 -d zend_extension=php-src/modules/opcache.so -d opcache.enable_cli=1 test.php
  Time (mean ± σ):      3.181 s ±  0.014 s    [User: 3.178 s, System: 0.003 s]
  Range (min … max):    3.165 s …  3.206 s    10 runs
 
Summary
  /tmp/unoptimized-commit1 -d zend_extension=php-src/modules/opcache.so -d opcache.enable_cli=1 test.php ran
    1.01 ± 0.02 times faster than sapi/cli/php -d zend_extension=php-src/modules/opcache.so -d opcache.enable_cli=1 test.php
    1.01 ± 0.01 times faster than /tmp/unoptimized -d zend_extension=php-src/modules/opcache.so -d opcache.enable_cli=1 test.php
```

-----------------------

Could it be that you accidentally compiled a debug build with compiler optimizations disabled or something like that?