@SakiTakamachi 2024-07-20
@TimWolla 

I measured these using gcc. I tried it with clang and got roughly the same results as yours.

```
// gcc
// before
Benchmark 1: php /mount/random/fromstr/t.php
  Time (mean ± σ):      5.876 s ±  0.214 s    [User: 5.871 s, System: 0.003 s]
  Range (min … max):    5.726 s …  6.285 s    10 runs

// after
Benchmark 1: php /mount/random/fromstr/t.php
  Time (mean ± σ):      4.833 s ±  0.170 s    [User: 4.829 s, System: 0.003 s]
  Range (min … max):    4.748 s …  5.291 s    10 runs



// clang
// before
Benchmark 1: php /mount/random/fromstr/t.php
  Time (mean ± σ):     10.864 s ±  0.180 s    [User: 10.859 s, System: 0.004 s]
  Range (min … max):   10.697 s … 11.162 s    10 runs

// after
Benchmark 1: php /mount/random/fromstr/t.php
  Time (mean ± σ):     10.595 s ±  0.239 s    [User: 10.590 s, System: 0.003 s]
  Range (min … max):   10.177 s … 11.000 s    10 runs
```

The configurations are as follows:
```
// gcc
./configure --disable-debug --disable-all --enable-bcmath --enable-tokenizer

// clang
./configure --disable-debug --disable-all --enable-bcmath --enable-tokenizer CC=clang
```

My gcc is a little old, so that might be the problem. I'll test it later in another environment.