@SakiTakamachi 2024-07-20
@TimWolla 
I tried it with gcc version 13.2.0, clang version 18.1.3 (ubuntu24.04).

The results are the same as before, with a significant difference in gcc, but almost no difference in clang. And clang is considerably slower than when compiled with gcc.

Could you please try it with gcc?